+++ arduinoProjects/gatewayFirmware/src/project/algo.h
@@ -85,7 +85,7 @@ class algo {
     // Helper for the date and time
     String getDateAndTime();
 
-    int calcNewOxy();
+    float calcNewOxy();
 };
 
 #endif
